+++ libqpdf/qpdf/RC4.hh
@@ -1,6 +1,8 @@
 #ifndef RC4_HH
 #define RC4_HH
 
+#include <stddef.h>
+
 class RC4
 {
   public:
@@ -8,7 +10,8 @@ class RC4
     RC4(unsigned char const* key_data, int key_len = -1);
 
     // out_data = 0 means to encrypt/decrypt in place
-    void process(unsigned char* in_data, int len, unsigned char* out_data = 0);
+    void process(unsigned char* in_data, size_t len,
+                 unsigned char* out_data = 0);
 
   private:
     class RC4Key
